GetSocialStat method parameters change from int to long

+++ CryptoCompare-Api/Clients/IOtherClient.cs
@@ -10,9 +10,9 @@ namespace CryptoCompare_Api.Clients
         ///     If called with the id of a cryptopian you just get data from our website that
         ///     is available to the public
         /// </summary>
-        /// <param name="id">The id of the coin/exchange you want social data for</param>
+        /// <param name="long">The id of the coin/exchange you want social data for</param>
         /// <returns></returns>
-        Task<Socialstat> GetSocialStat(int id);
+        Task<Socialstat> GetSocialStat(long id);
 
         /// <summary>
         ///     Get general info for all the coins available on the website.
